requirements.txt
PyYAML==6.0.2
netboxlabs-diode-sdk==0.4.2
pynetbox==7.4.1